Conflicts:(Keep linux4.4-rk3288-tinker_board link file.)
	rk3288/BoardConfig.mk

+++ oem/oem_uvcc/aicamera.sh
@@ -1,17 +1,55 @@
 #!/bin/sh
 #
 
+TRY_CNT=0
+check_uvc_suspend()
+{
+  if [ -e /tmp/uvc_goto_suspend ];then
+     echo "uvc go to suspend now"
+     killall ispserver
+     killall aiserver
+     sleep 2
+     if [ -e /tmp/uvc_goto_suspend ];then
+       rm /tmp/uvc_goto_suspend -rf
+       echo mem > /sys/power/state
+     fi
+  fi
+}
+
+check_uvc_buffer()
+{
+  if [ "$TRY_CNT" -gt 0 ];then
+     let TRY_CNT=TRY_CNT-1
+     #echo "++++++++TRY_CNT:$TRY_CNT"
+  fi
+  if [ "$TRY_CNT" -gt 10 ];then
+     echo "+++check_uvc_buffer recovery fail,reboot to recovery now+++"
+     reboot &
+  fi
+  if [ -e /tmp/uvc_camera_no_buf ];then
+     let TRY_CNT=TRY_CNT+10
+     echo "uvc no buf to send 200 frames,try to recovery isp time,timeout:$TRY_CNT"
+     killall ispserver
+     killall aiserver
+     rm /tmp/uvc_camera_no_buf -rf
+  fi
+}
 check_alive()
 {
+  if [[ ! -f "/oem/usr/bin/$1"  && ! -f "/usr/bin/$1" ]]; then
+   return 1
+  fi
   PID=`busybox ps |grep $1 |grep -v grep | wc -l`
   if [ $PID -le 0 ];then
      if [ "$1"x == "uvc_app"x ];then
        echo " uvc app die ,restart it and usb reprobe !!!"
-       echo none > /sys/kernel/config/usb_gadget/rockchip/UDC
        sleep 1
        rm -rf /sys/kernel/config/usb_gadget/rockchip/configs/b.1/f*
-       echo ffd00000.dwc3  > /sys/bus/platform/drivers/dwc3/unbind
-       echo ffd00000.dwc3  > /sys/bus/platform/drivers/dwc3/bind
+       echo none > /sys/kernel/config/usb_gadget/rockchip/UDC
+       rmdir /sys/kernel/config/usb_gadget/rockchip/functions/rndis.gs0
+       UDC=`ls /sys/class/udc/| awk '{print $1}'`
+       echo $UDC  > /sys/bus/platform/drivers/dwc3/unbind
+       echo $UDC  > /sys/bus/platform/drivers/dwc3/bind
        /oem/usb_config.sh rndis off #disable adb
        usb_irq_set
        uvc_app &
@@ -23,6 +61,8 @@ check_alive()
             echo "aiserver is die,tell uvc to recovery"
             killall -3 uvc_app
             aiserver &
+            sleep .5
+            killall -10 smart_display_service
          else
             $1 &
          fi
@@ -48,20 +88,26 @@ usb_irq_set()
   echo "usb irq:$usbirq"
   echo 1 > /proc/irq/$usbirq/smp_affinity_list
 }
-ulimit -c unlimited
+#ulimit -c unlimited
 dbserver &
 ispserver -n &
 stop_unused_daemon
+#uac_app &
 /oem/usb_config.sh rndis
 usb_irq_set
 uvc_app &
 aiserver &
+sleep .5
+smart_display_service &
 while true
 do
   check_alive dbserver
   check_alive ispserver
   check_alive uvc_app
+#  check_alive uac_app
   check_alive aiserver
+#  check_uvc_buffer
+#  check_uvc_suspend
   sleep 2
   check_alive smart_display_service
 done
